+++ CHANGELOG.md
@@ -21,6 +21,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - Changed OAuth authorization code lifetime to 5 minutes.
 - Log `redirect_uri` and `client_id` parameters when OAuth access token is requested.
 - Save media digests to database.
+- Overwrite cached activity if attributed object has same ID.
 
 ### Fixed